refactor: tsx, emotion 파일 분리
