+++ src/components/main/section/hot-potfolio/index.tsx
@@ -1,9 +1,16 @@
 /* eslint-disable @typescript-eslint/no-use-before-define */
 import React from "react";
-import styled from "@emotion/styled";
-import { GREY } from "../../../../constants/colors";
 import PortfolioCard from "./portfolio-card";
 import users from "./fakeUser";
+import {
+  Container,
+  ArticleWrapper,
+  Title,
+  Category,
+  CategoryWrapper,
+  CardWrapper,
+  StyledButton,
+} from "./index.styles";
 
 export default function MainArticle1() {
   const CategoryList = ["전체보기", "음악", "연기", "연출", "모델", "+"];
@@ -26,80 +33,3 @@ export default function MainArticle1() {
     </Container>
   );
 }
-
-const Title = styled.div`
-  width: 100%;
-  font-size: 1.7rem;
-  font-weight: 700;
-  text-align: start;
-  margin-bottom: 2rem;
-`;
-const Container = styled.div`
-  width: 100%;
-  /* height: 70rem; */
-  display: flex;
-  justify-content: center;
-  margin-top: 8rem;
-`;
-
-const ArticleWrapper = styled.div`
-  width: 68%;
-  height: 100%;
-  display: flex;
-  flex-wrap: wrap;
-  flex-direction: column;
-`;
-
-const CategoryWrapper = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  height: 18%;
-  & :last-child {
-    width: 8%;
-    font-size: 2rem;
-  }
-  padding-bottom: 2.5rem;
-`;
-
-const CardWrapper = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  height: 30%;
-  padding-bottom: 1rem;
-`;
-
-const Category = styled.button`
-  background-color: ${GREY[700]};
-  width: 17%;
-  height: 9rem;
-  margin: 0.5rem 0;
-  border-radius: 0.5rem;
-  border: none;
-  text-align: center;
-  font-weight: 700;
-  font-size: 1.5rem;
-  color: white;
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-const StyledButton = styled.button`
-  margin-top: 5rem;
-  width: 100%;
-  height: 6rem;
-  border: none;
-  background-color: ${GREY[800]};
-  border-radius: 6rem;
-  color: white;
-  font-weight: 700;
-  font-size: 1.7rem;
-  text-align: center;
-  &:hover {
-    cursor: pointer;
-  }
-`;
